Adhere to openPMD requirements before flushing
For a given particle species, the openPMD API requires that required records (such as "position", "positionOffset") and their contained components be defined (not necessarily written). Make sure to define all such datasets before issuing the first flush.

Maybe open an issue in the openPMD API to allow for a more flexible usage.